Remove old Alt Manager
