+++ src/main/java/cope/cosmos/client/Cosmos.java
@@ -122,7 +122,7 @@ public class Cosmos {
         // sets up the social manager
         socialManager = new SocialManager();
         // Grouping this with socials, move it if it should be somewhere else
-        AlttManager.loadAlts();
+        AltManager.loadAlts();
         managers.add(socialManager);
         progressManager.step("Setting up Social Manager");
