+++ src/main/java/cope/cosmos/client/manager/managers/AltManager.java
@@ -1,73 +1,115 @@
 package cope.cosmos.client.manager.managers;
 
-import com.mojang.authlib.Agent;
-import com.mojang.authlib.yggdrasil.YggdrasilAuthenticationService;
-import com.mojang.authlib.yggdrasil.YggdrasilUserAuthentication;
-import cope.cosmos.asm.mixins.accessor.IMinecraft;
-import cope.cosmos.client.ui.altgui.AltEntry;
-import cope.cosmos.client.manager.Manager;
-import net.minecraft.util.Session;
+import com.moandjiezana.toml.Toml;
+import cope.cosmos.client.ui.altmanager.Alt;
+import cope.cosmos.client.ui.altmanager.AltEntry;
+import cope.cosmos.client.ui.altmanager.AltManagerGUI;
 
-import java.net.Proxy;
+import java.io.*;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
 
 /**
- * @author bon55
- * @since 05/06/2021
+ * @author Wolfsurge, linustouchtips
  */
-public class AltManager extends Manager {
+public class AltManager {
 
-	// list of all alternate accounts
-	private static final List<AltEntry> alts = new ArrayList<>();
+    // List of alt entries
+    private static final List<AltEntry> altEntries = new ArrayList<>();
 
-	public AltManager() {
-		super("AltManager", "Manages alternate accounts for easy login");
-	}
+    // The client directory
+    private static final File mainDirectory = new File("cosmos");
 
-	/**
-	 * Logs into a given Minecraft account
-	 * @param email Email to the Minecraft account
-	 * @param password Password to the Minecraft account
-	 * @param setSession If to attempt to set the current session as the new login
-	 * @return The authentication state of the Minecraft account
-	 */
-	public static YggdrasilUserAuthentication logIn(String email, String password, boolean setSession) {
-		// authentication to mc servers
-		YggdrasilUserAuthentication auth = (YggdrasilUserAuthentication) new YggdrasilAuthenticationService(Proxy.NO_PROXY, "").createUserAuthentication(Agent.MINECRAFT);
+    public AltManager() {
 
-		// update auth info
-		auth.setUsername(email);
-		auth.setPassword(password);
+    }
 
-		new Thread(() -> {
-			try {
-				// attempt to login
-				auth.logIn();
+    /**
+     * Saves the alt accounts to alts.toml
+     */
+    public static void saveAlts() {
+        try {
+            // File writer
+            OutputStreamWriter altOutputStreamWriter = new OutputStreamWriter(new FileOutputStream(mainDirectory.getName() + "/alts.toml"), StandardCharsets.UTF_8);
 
-				// attempt to set the session
-				if (setSession) {
-					Session session = new Session(auth.getSelectedProfile().getName(), auth.getSelectedProfile().getId().toString(), auth.getAuthenticatedToken(), "mojang");
+            // Output
+            StringBuilder output = new StringBuilder();
 
-					// session null?
-					if (session != null) {
-						((IMinecraft) mc).setSession(session);
-					}
-				}
-			} catch (Exception exception) {
-				exception.printStackTrace();
-			}
-		}).start();
+            try {
+                output.append("[Alts]").append("\r\n");
 
-		// return authentication
-		return auth;
-	}
+                output.append("Alts").append(" = ").append("[");
+                // Add the alt's info, in the order: email:password:type
+                for(AltEntry altEntry : getAltEntries()) {
+                    output.append('"').append(altEntry.getAlt().getLogin()).append(":").append(altEntry.getAlt().getPassword()).append(":").append(altEntry.getAlt().getAltType().name()).append('"').append(",");
+                }
+                output.append("]").append("\r\n");
+            } catch (Exception exception) {
+                exception.printStackTrace();
+            }
 
-	/**
-	 * Gets a list of all alt accounts
-	 * @return List of all alt accounts
-	 */
-	public static List<AltEntry> getAlts() {
-		return alts;
-	}
+            // Write the info
+            altOutputStreamWriter.write(output.toString());
+            altOutputStreamWriter.close();
+        }
+        catch (IOException exception) {
+            exception.printStackTrace();
+        }
+    }
+
+    /**
+     * Loads the alts from alts.toml
+     */
+    public static void loadAlts() {
+        try {
+            // The stream from alts file
+            InputStream inputStream = Files.newInputStream(Paths.get(mainDirectory.getName() + "/alts.toml"));
+
+            // Input TOML
+            Toml input = new Toml().read(inputStream);
+
+            if(input != null) {
+                try {
+                    if(input.getList("Alts.Alts") != null) {
+                        input.getList("Alts.Alts").forEach(object -> {
+                            // 0 = Email, 1 = Password, 2 = Type
+                            String[] altInfo = ((String) object).split(":");
+
+                            Alt.AltType type = null;
+                            if(altInfo[2].equalsIgnoreCase("Microsoft")) type = Alt.AltType.Microsoft;
+                            else if(altInfo[2].equalsIgnoreCase("Mojang")) type = Alt.AltType.Mojang;
+                            else if(altInfo[2].equalsIgnoreCase("Cracked")) type = Alt.AltType.Cracked;
+
+                            // Loading alt
+                            Alt alt = new Alt(altInfo[0], altInfo[1], type);
+
+                            // Add alt to list
+                            getAltEntries().add(new AltEntry(alt, AltManagerGUI.altEntryOffset));
+                            // Add to offset
+                            AltManagerGUI.altEntryOffset += 32;
+                        });
+                    }
+                }
+                catch (Exception exception) {
+                    exception.printStackTrace();
+                }
+            }
+        }
+        catch (IOException exception) {
+            exception.printStackTrace();
+        }
+    }
+
+    /**
+     * Gets the alt entries
+     * @return The alt entries
+     */
+    public static List<AltEntry> getAltEntries() {
+        return altEntries;
+    }
+
 }
